Implment buy meal plan + subscription

+++ FE/NomsNoms/src/app/components/admin/user-management/user-management.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, ElementRef, OnInit, Renderer2, ViewChild} from '@angular/core';
 import {User} from "../../../_model/user.model";
 import {Subject} from "rxjs";
 import {UserService} from "../../../_services/user.service";
@@ -16,8 +16,10 @@ export class UserManagementComponent implements OnInit{
     pagingType: 'full_numbers'
   };
   dtTrigger: Subject<any> = new Subject<any>();
+  @ViewChild('userTable') userTable: ElementRef;
   constructor(private uService: UserService,
-              private toastr: ToastrService) {
+              private toastr: ToastrService,
+              private renderer: Renderer2) {
   }
   ngOnInit() {
     this.loadUsers();
@@ -32,11 +34,48 @@ export class UserManagementComponent implements OnInit{
   }
 
   deleteUser(user: UserAdmin) {
-    console.log(user);
+    if (!user) return;
     this.uService.deleteUser(user).subscribe({
       next: _ => {
-        this.loadUsers();
-        this.toastr.success("Vô hiệu hóa người dùng thành công");
+        if (this.users)
+          var user1 = this.users.find(x => x.id == user.id);
+        if (user1) {
+          user1.status = 0;
+
+          const table = this.userTable.nativeElement;
+          const row = table.querySelector(`[data-user-id="${user.id}"]`);
+
+          if (row) {
+            const cell = row.cells[3];
+
+            // update the cell content
+            this.renderer.setProperty(cell, 'textContent', "Vô hiệu hóa");
+          }
+          this.toastr.success("Vô hiệu hóa người dùng thành công");
+        }
+      }
+    });
+  }
+  enableUser(user: UserAdmin) {
+    console.log(user);
+    this.uService.enable(user).subscribe({
+      next: _ => {
+        if (this.users)
+          var user1 = this.users.find(x => x.id == user.id);
+        if (user1) {
+          user1.status = 1;
+
+          const table = this.userTable.nativeElement;
+          const row = table.querySelector(`[data-user-id="${user.id}"]`);
+
+          if (row) {
+            const cell = row.cells[3];
+
+            // update the cell content
+            this.renderer.setProperty(cell, 'textContent', "Hoạt động");
+          }
+          this.toastr.success("Mở khóa người dùng thành công");
+        }
       }
     });
   }
